requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.7.0
groq==0.4.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
huggingface-hub==0.20.0
